// web/src/api/generated/CourseSelector.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Attribute } from "./Attribute";
import type { CourseCode } from "./CourseCode";
import type { Subject } from "./Subject";

/**
 * One way to name courses.
 */
export type CourseSelector = { "kind": "code", 
/**
 * The code.
 */
code: CourseCode, } | { "kind": "subject", 
/**
 * The subject.
 */
subject: Subject, } | { "kind": "number_range", 
/**
 * The subject, or any.
 */
subject?: Subject | null, 
/**
 * Lowest number.
 */
low: number, 
/**
 * Highest number.
 */
high: number, } | { "kind": "attribute", 
/**
 * The attribute.
 */
attribute: Attribute, };